machine_learning/requirements.txt
transformers[torch]
datasets
flask
protobuf
tiktoken
python-dotenv
trl
sentencepiece>=0.1.99